pyproject.toml
[tool.poetry]
name = "currency-converter-app"
version = "0.1.0"
description = "Arabic RTL Currency Conversion Chatbot using Flask and Gemini API"
authors = ["Mohamed Fakhry <mohamedfakhrysmile@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = { version = "^2.3.3", extras = ["async"] }
aiohttp = "^3.9.3"
nest-asyncio = "^1.5.8"
python-dotenv = "^1.0.0"
google-generativeai = "^0.8.4"
flask-cors = "^4.0.0"
gunicorn = "^21.2.0"
loguru = "^0.7.2"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.packages]]
include = "src"